abstract out toggleAllInputs
